fix(test): move test and refactor them
